Remove red crash message from test output
Riemann.ConnectionTest:26 was outputting the following message in the
test results when the tcp connection went down:

10:24:22.856 [error] GenServer #PID<0.176.0> terminating
** (stop) :tcp_closed
Last message: {:tcp_closed, #Port<0.5458>}
State: %Riemann.Connection.State{from: nil, tcp: nil}

This is an expected log message, but doesn't promote a lot of confidence
in the test suite when there is red in the output but all of the tests
say the pass.

This wraps the shutting down of the test server in a capture_log
statement which makes sure the expected :tcp_closed message is logged.