Read :event_host setting at runtime
Currently, the :event_host setting is read at compile-time and can't be
changed at runtime via settings in sys.config. This changes the code so
the :event_host setting is read at runtime instead of compile time.

I also moved the logic for building an event into Riemann.Helpers.Event
instead of creating a bunch of private functions that pollute the module
that `uses` this module.